Modified streamlit app code
